Unify unit test naming: use '*It.java' instead of '*IT.java'.
