+++ rider-spring/pom.xml
@@ -83,7 +83,7 @@
                 <version>2.19.1</version>
                 <configuration>
                     <includes>
-                        <include>**/*IT.java</include>
+                        <include>**/*It.java</include>
                         <include>**/*Test.java</include>
                     </includes>
                     <systemPropertyVariables>
